Made it so there is one CommentSection being applied rather than each feed having their own.
